Functional parts moved from main component to subcomponent.
Naming corrected for subcomponent to follow guidelines.
