+++ assets/stylesheets/_components.scss
@@ -287,6 +287,7 @@
 @import 'my-sites/stats/stats-tabs/style';
 @import 'my-sites/theme/style';
 @import 'my-sites/theme/theme-download-card/style';
+@import 'my-sites/theme/related-themes-card/style';
 @import 'my-sites/themes/current-theme/style';
 @import 'my-sites/themes/style';
 @import 'my-sites/themes/themes-search-card/style';
