+++ client/my-sites/theme/main.jsx
@@ -1,4 +1,4 @@
-/** @ssr-ready **/
+///** @ssr-ready **/
 
 /* eslint-disable react/no-danger  */
 // FIXME!!!^ we want to ensure we have sanitised data…
@@ -17,6 +17,7 @@ import Main from 'components/main';
 import HeaderCake from 'components/header-cake';
 import SectionHeader from 'components/section-header';
 import ThemeDownloadCard from './theme-download-card';
+import ThemeRelatedThemesCard from './related-themes-card';
 import Button from 'components/button';
 import SectionNav from 'components/section-nav';
 import NavTabs from 'components/section-nav/tabs';
@@ -292,72 +293,8 @@ const ThemeSheet = React.createClass( {
 	},
 
 	renderRelatedThemes() {
-		var actionLabel = i18n.translate( 'Pick' ),
-		themes = this.getRelatedThemes().map( function( theme ) {
-				return {
-					id: theme.slug,
-					screenshot: 'https://i1.wp.com/s0.wp.com/wp-content/themes/pub/' + theme.slug + '/screenshot.png'
-				};
-			} );
 
-		const related_text = i18n.translate( 'See all {{a}}BUSINESS{{/a}} themes.', {
-				components: {
-					a: <a href='https://wordpress.com'/>
-				}
-			} );
-
-		return (
-			<div>
-				<SectionHeader label={ i18n.translate( 'You might also like' ) } />
-					<ul className="themes__sheet-related-themes">
-						{ themes.map( function( theme ) {
-							return (
-								<li key={ theme.id }>
-									<Card className="themes__sheet-related-themes-card">
-										<a href={ getDetailsUrl( theme ) }>
-											<img src={ theme.screenshot + '?w=' + '660' }/>
-										</a>
-									</Card>
-								</li>
-							);
-						} ) }
-					</ul>
-					<div className="themes__sheet-related-themes-link">
-						<p>{ related_text }</p>
-					</div>
-			</div>
-		);
-	},
-
-	getRelatedThemes() {
-		var themes = [
-			{ name: 'Twenty Sixteen', slug: 'twentysixteen' },
-			{ name: 'Rowling', slug: 'rowling' },
-			{ name: 'Hemingway Rewritten', slug: 'hemingway-rewritten' },
-			{ name: 'Gazette', slug: 'gazette' },
-			{ name: 'Ecto', slug: 'ecto' },
-			{ name: 'Isola', slug: 'isola' },
-			{ name: 'Edin', slug: 'edin' },
-			{ name: 'Sela', slug: 'sela' },
-			{ name: 'Pique', slug: 'pique' },
-			{ name: 'Harmonic', slug: 'harmonic' },
-		];
-
-		var selectedThemes = [];
-
-		while( selectedThemes.length < 2 ) {
-			var randomThemeIndex = Math.floor(Math.random() * themes.length );
-			var theme = themes.splice( randomThemeIndex, 1 )[0];
-			// Check if it is not current theme, if it is than discard
-			if( this.props.name === theme.name ) {
-				continue;
-			} else {
-				selectedThemes.push( theme );
-			}
-
-		}
-
-		return selectedThemes;
+		return <ThemeRelatedThemesCard />;
 	},
 
 	renderSheet() {
